Move all of the Fiber logic to FiberManager
* Simplify by using instance methods.
