+++ src/main/java/org/truffleruby/language/SafepointManager.java
@@ -21,7 +21,7 @@ import org.truffleruby.Layouts;
 import org.truffleruby.Log;
 import org.truffleruby.RubyContext;
 import org.truffleruby.core.InterruptMode;
-import org.truffleruby.core.fiber.FiberNodes;
+import org.truffleruby.core.fiber.FiberManager;
 import org.truffleruby.core.thread.ThreadStatus;
 import org.truffleruby.platform.signal.Signal;
 
@@ -301,7 +301,8 @@ public class SafepointManager {
     public String getSafepointDebugInfo() {
         final Thread[] threads = new Thread[Thread.activeCount() + 1024];
         final int threadsCount = Thread.enumerate(threads);
-        final long appearRunning = Arrays.stream(threads).limit(threadsCount).filter((t) -> t.getName().startsWith(FiberNodes.NAME_PREFIX)).count();
+        final long appearRunning = Arrays.stream(threads).limit(threadsCount).filter(
+                t -> t.getName().startsWith(FiberManager.NAME_PREFIX)).count();
         return String.format("safepoints: %d known threads, %d registered with phaser, %d arrived, %d appear to be running",
                 runningThreads.size(), phaser.getRegisteredParties(), phaser.getArrivedParties(), appearRunning);
     }
